This commit includes modification of the DialogueManager along with a test scene with the existing NPC
The modification of the DialogueManager was required so that we can get
the "interact" key to work properly with dialogue.

In addition, hooked up the DialogueManager's on_dialogue_end_signal to
prevent the character from moving when choosing responses from the
dialogue.

Added a new test scene with the NPC existing and working with the
dialogue responses.